projects/my-cdk/src/avatar/avatar.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { AvatarComponent } from './avatar.component';

describe('AvatarComponent', () => {
  let component: AvatarComponent;
  let fixture: ComponentFixture<AvatarComponent>;

  beforeEach(() => {
    TestBed.configureTestingModule({
      declarations: [AvatarComponent],
    }).compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(AvatarComponent);
    component = fixture.componentInstance;
  });

  it('should create', () => {
    fixture.detectChanges();
    expect(component).toBeTruthy();
  });

  it('should display current url', () => {
    component.url = 'test.jpg';
    fixture.detectChanges();
    const compiled = fixture.debugElement.nativeElement;
    expect(compiled.querySelector('img').attributes['src'].value).toEqual('test.jpg');
    expect(compiled.querySelector('svg')).toBeFalsy();
  });

  it('should display placeholder', () => {
    component.url = undefined;
    fixture.detectChanges();
    const compiled = fixture.debugElement.nativeElement;
    expect(compiled.querySelector('svg')).toBeTruthy();
    expect(compiled.querySelector('img')).toBeFalsy();
  });
});
